+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-[Unreleased]
+[5.15.0] - 2020-04-24
 ----------------------
 ##### Added
 - Support for multiple setvalue/xforms-value-changed actions under the same form control.